# Move CSS styling to Javascript