--- a/src/cells/gates.mjs
+++ b/src/cells/gates.mjs
@@ -14,8 +14,8 @@ const xor_arc_path = "M6.8 2.8L10 6.7S14.2 12 14.2 20 10 33.3 10 33.3l-3.2 3.9H1
 const xor_arc_path_markup = {
     tagName: 'path',
     attributes: {
-        fill: "#000",
-        d: xor_arc_path
+        'fill': "#000",
+        'd': xor_arc_path
     }
 };
 
@@ -23,9 +23,11 @@ const neg_markup = {
     tagName: 'circle',
     className: 'body',
     attributes: {
-        cx: 56,
-        cy: 20,
-        r: 4
+        'stroke': "#000",
+        'stroke-width': '2px',
+        'cx': 56,
+        'cy': 20,
+        'r': 4
     }
 };
 
@@ -37,14 +39,15 @@ export const GateSVG = Gate.define('GateSVG', {
     size: { width: 60, height: 40 },
     ports: {
         groups: {
-            'in': { position: { name: 'left', args: { dx: 20 } }, attrs: { 'line.wire': { x2: -40 }, port: { refX: -40 } }, z: -1 },
-            'out': { position: { name: 'right', args: { dx: -20 } }, attrs: { 'line.wire': { x2: 40 }, port: { refX: 40 } }, z: -1 }
+            'in': { position: { name: 'left', args: { dx: 20 } }, attrs: { wire: { x2: -40 }, port: { refX: -40 } }, z: -1 },
+            'out': { position: { name: 'right', args: { dx: -20 } }, attrs: { wire: { x2: 40 }, port: { refX: 40 } }, z: -1 }
         }
     }
 }, {
     markup: Gate.prototype.markup.concat([{
             tagName: 'path',
-            className: 'body gate'
+            className: 'body gate',
+            selector: 'body'
         }
     ]),
     gateParams: Gate.prototype.gateParams.concat(['bits'])
@@ -104,7 +107,7 @@ export const GateReduce = GateSVG.define('GateReduce', {}, {
 
 // Repeater (buffer) gate model
 export const Repeater = Gate11.define('Repeater', {
-    attrs: { 'path.gate': { d: buf_path }}
+    attrs: { body: { d: buf_path }}
 }, {
     operation: function(data) {
         return { out: data.in };
@@ -114,7 +117,7 @@ export const RepeaterView = GateView;
 
 // Not gate model
 export const Not = Gate11.define('Not', {
-    attrs: { 'path.gate': { d: buf_path }}
+    attrs: { body: { d: buf_path }}
 }, {
     operation: function(data) {
         return { out: data.in.not() };
@@ -125,7 +128,7 @@ export const NotView = GateView;
 
 // Or gate model
 export const Or = Gate21.define('Or', {
-    attrs: { 'path.gate': { d: or_path }}
+    attrs: { body: { d: or_path }}
 }, {
     operation: function(data) {
         return { out: data.in1.or(data.in2) };
@@ -135,7 +138,7 @@ export const OrView = GateView;
 
 // And gate model
 export const And = Gate21.define('And', {
-    attrs: { 'path.gate': { d: and_path }}
+    attrs: { body: { d: and_path }}
 }, {
     operation: function(data) {
         return { out: data.in1.and(data.in2) };
@@ -145,7 +148,7 @@ export const AndView = GateView;
 
 // Nor gate model
 export const Nor = Gate21.define('Nor', {
-    attrs: { 'path.gate': { d: or_path }}
+    attrs: { body: { d: or_path }}
 }, {
     operation: function(data) {
         return { out: data.in1.nor(data.in2) };
@@ -156,7 +159,7 @@ export const NorView = GateView;
 
 // Nand gate model
 export const Nand = Gate21.define('Nand', {
-    attrs: { 'path.gate': { d: and_path }}
+    attrs: { body: { d: and_path }}
 }, {
     operation: function(data) {
         return { out: data.in1.nand(data.in2) };
@@ -167,7 +170,7 @@ export const NandView = GateView;
 
 // Xor gate model
 export const Xor = Gate21.define('Xor', {
-    attrs: { 'path.gate': { d: or_path }}
+    attrs: { body: { d: or_path }}
 }, {
     operation: function(data) {
         return { out: data.in1.xor(data.in2) };
@@ -178,7 +181,7 @@ export const XorView = GateView;
 
 // Xnor gate model
 export const Xnor = Gate21.define('Xnor', {
-    attrs: { 'path.gate': { d: or_path }}
+    attrs: { body: { d: or_path }}
 }, {
     operation: function(data) {
         return { out: data.in1.xnor(data.in2) };
@@ -189,7 +192,7 @@ export const XnorView = GateView;
 
 // Reducing Or gate model
 export const OrReduce = GateReduce.define('OrReduce', {
-    attrs: { 'path.gate': { d: or_path }}
+    attrs: { body: { d: or_path }}
 }, {
     operation: function(data) {
         return { out: data.in.reduceOr() };
@@ -199,7 +202,7 @@ export const OrReduceView = GateView;
 
 // Reducing Nor gate model
 export const NorReduce = GateReduce.define('NorReduce', {
-    attrs: { 'path.gate': { d: or_path }}
+    attrs: { body: { d: or_path }}
 }, {
     operation: function(data) {
         return { out: data.in.reduceNor() };
@@ -210,7 +213,7 @@ export const NorReduceView = GateView;
 
 // Reducing And gate model
 export const AndReduce = GateReduce.define('AndReduce', {
-    attrs: { 'path.gate': { d: and_path }}
+    attrs: { body: { d: and_path }}
 }, {
     operation: function(data) {
         return { out: data.in.reduceAnd() };
@@ -220,7 +223,7 @@ export const AndReduceView = GateView;
 
 // Reducing Nand gate model
 export const NandReduce = GateReduce.define('NandReduce', {
-    attrs: { 'path.gate': { d: and_path }}
+    attrs: { body: { d: and_path }}
 }, {
     operation: function(data) {
         return { out: data.in.reduceNand() };
@@ -231,7 +234,7 @@ export const NandReduceView = GateView;
 
 // Reducing Xor gate model
 export const XorReduce = GateReduce.define('XorReduce', {
-    attrs: { 'path.gate': { d: or_path }}
+    attrs: { body: { d: or_path }}
 }, {
     operation: function(data) {
         return { out: data.in.reduceXor() };
@@ -242,7 +245,7 @@ export const XorReduceView = GateView;
 
 // Reducing Xnor gate model
 export const XnorReduce = GateReduce.define('XnorReduce', {
-    attrs: { 'path.gate': { d: or_path }}
+    attrs: { body: { d: or_path }}
 }, {
     operation: function(data) {
         return { out: data.in.reduceXnor() };