This PR fixes #23 in two steps:
- switching from classNames to [jointjs selectors](https://resources.jointjs.com/tutorial/custom-elements#markup) for default `attrs` styling ("Selectors are important for the targeting of subelements by element attributes. Although providing a selector to identify a subelement is not strictly required, it makes JointJS noticeably faster because it can avoid querying the DOM."). For my experiments this change did not produce any noticeable performance boost, but at least it ran as fast as before.
- Instead of toggling corresponding classes in the `View`s, temporary attributes (such as "high", "low" or "bus") are directly set on the element for each value change. The temporary attributes might be adapted by overwriting `this.attrs` inside the View.

As already pointed out in #23, this means the SVG is now fully defined in Javascript and no further CSS styling is required.

Benchmarking of the changes gave the following result on the Horner example (with open `multiplier` of last `adder_multiplier` in the chain):
| Input change | before | after |
|--|--|--|
| `xxx` -> `00f` | 22s | 21s |
| `00f` -> `x` | 9s | 9s |

The changes did not noticeably alter the performance, most of the time indeed seems to be spent by the model calculation and not by updating the views.